Refactor cache response classes
- Move cache response classes and rename.
- Updated new Headers response object to be able to handle items
